requirements.txt
discord.py
google-cloud-automl